Add config parameter to applyValueOperator()
