+++ src/Glpi/Form/Condition/ConditionHandler/AbstractDateTimeConditionHandler.php
@@ -34,6 +34,7 @@
 
 namespace Glpi\Form\Condition\ConditionHandler;
 
+use Glpi\DBAL\JsonFieldInterface;
 use Glpi\Form\Condition\ValueOperator;
 use Override;
 
@@ -63,6 +64,7 @@ abstract class AbstractDateTimeConditionHandler implements ConditionHandlerInter
         mixed $a,
         ValueOperator $operator,
         mixed $b,
+        ?JsonFieldInterface $config,
     ): bool {
         // Date can be compared as simple strings.
         $a = strtolower(strval($a));
