+++ src/Glpi/Form/Condition/ConditionHandler/ActorConditionHandler.php
@@ -34,6 +34,7 @@
 
 namespace Glpi\Form\Condition\ConditionHandler;
 
+use Glpi\DBAL\JsonFieldInterface;
 use Glpi\Form\Condition\ConditionData;
 use Glpi\Form\Condition\ValueOperator;
 use Glpi\Form\QuestionType\AbstractQuestionTypeActors;
@@ -75,6 +76,7 @@ class ActorConditionHandler implements ConditionHandlerInterface
         mixed $a,
         ValueOperator $operator,
         mixed $b,
+        ?JsonFieldInterface $config,
     ): bool {
         return $this->applyArrayValueOperator($a, $operator, $b);
     }
