+++ src/Glpi/Form/Condition/ConditionHandler/ItemAsTextConditionHandler.php
@@ -34,6 +34,7 @@
 
 namespace Glpi\Form\Condition\ConditionHandler;
 
+use Glpi\DBAL\JsonFieldInterface;
 use Glpi\Form\Condition\ConditionData;
 use Glpi\Form\Condition\ValueOperator;
 use Override;
@@ -73,6 +74,7 @@ final class ItemAsTextConditionHandler implements ConditionHandlerInterface
         mixed $a,
         ValueOperator $operator,
         mixed $b,
+        ?JsonFieldInterface $config,
     ): bool {
         // $a is the submitted answer
         if (!is_array($a) || !isset($a['items_id'])) {
