+++ src/Glpi/Form/Condition/ConditionHandler/MultipleChoiceFromValuesConditionHandler.php
@@ -34,6 +34,7 @@
 
 namespace Glpi\Form\Condition\ConditionHandler;
 
+use Glpi\DBAL\JsonFieldInterface;
 use Glpi\Form\Condition\ConditionData;
 use Glpi\Form\Condition\ValueOperator;
 use Glpi\Form\Migration\ConditionHandlerDataConverterInterface;
@@ -74,6 +75,7 @@ final class MultipleChoiceFromValuesConditionHandler implements
         mixed $a,
         ValueOperator $operator,
         mixed $b,
+        ?JsonFieldInterface $config,
     ): bool {
         return $this->applyArrayValueOperator($a, $operator, $b);
     }
