+++ src/Glpi/Form/Condition/ConditionHandler/RichTextConditionHandler.php
@@ -34,6 +34,7 @@
 
 namespace Glpi\Form\Condition\ConditionHandler;
 
+use Glpi\DBAL\JsonFieldInterface;
 use Glpi\Form\Condition\ValueOperator;
 use Override;
 
@@ -44,11 +45,12 @@ final class RichTextConditionHandler extends StringConditionHandler implements C
         mixed $a,
         ValueOperator $operator,
         mixed $b,
+        ?JsonFieldInterface $config,
     ): bool {
         // Remove HTML tags.
         $a = strip_tags(strval($a));
         $b = strip_tags(strval($b));
 
-        return parent::applyValueOperator($a, $operator, $b);
+        return parent::applyValueOperator($a, $operator, $b, $config);
     }
 }
